Rename command package
